requirements.txt
faiss-cpu==1.13.2
PyMuPDF==1.26.7
langchain-openai==1.1.7
dotenv
langgraph==1.0.5

fastapi==0.128.0
uvicorn[standard]==0.27.0
Jinja2==3.1.6

google-cloud-translate==3.23.0

python-multipart==0.0.21

itsdangerous==2.2.0

python-docx==1.1.0